+++ apps/backend/src/modules/points-of-interest/libs/validation-schemas/validation-schemas.ts
@@ -1,0 +1,4 @@
+export {
+	pointOfInterestCreateValidationSchema,
+	pointOfInterestUpdateValidationSchema,
+} from "@smartscapes/shared";