+++ apps/backend/src/modules/points-of-interest/points-of-interest.model.ts
@@ -1,0 +1,14 @@
+import {
+	AbstractModel,
+	DatabaseTableName,
+} from "~/libs/modules/database/database.js";
+
+class PointsOfInterestModel extends AbstractModel {
+	public static override get tableName(): string {
+		return DatabaseTableName.POINTS_OF_INTEREST;
+	}
+
+	public name!: string;
+}
+
+export { PointsOfInterestModel };