+++ apps/frontend/src/modules/auth/slices/actions.ts
@@ -5,6 +5,8 @@ import { type APIResponse, type AsyncThunkConfig } from "~/libs/types/types.js";
 
 import {
 	type UserAuthResponseDto,
+	type UserSignInRequestDto,
+	type UserSignInResponseDto,
 	type UserSignUpRequestDto,
 	type UserSignUpResponseDto,
 } from "../libs/types/types.js";
@@ -42,4 +44,16 @@ const signUp = createAsyncThunk<
 	return user;
 });
 
-export { getAuthenticatedUser, signUp };
+const signIn = createAsyncThunk<
+	APIResponse<UserSignInResponseDto>,
+	UserSignInRequestDto,
+	AsyncThunkConfig
+>(`${sliceName}/sign-in`, async (payload, { extra }) => {
+	const { authApi } = extra;
+
+	const user = await authApi.signIn(payload);
+
+	return user;
+});
+
+export { getAuthenticatedUser, signIn, signUp };