+++ apps/frontend/src/modules/auth/slices/auth.slice.ts
@@ -4,7 +4,7 @@ import { type UserAuthResponseDto } from "@smartscapes/shared";
 import { DataStatus } from "~/libs/enums/enums.js";
 import { type ValueOf } from "~/libs/types/types.js";
 
-import { getAuthenticatedUser, signUp } from "./actions.js";
+import { getAuthenticatedUser, signIn, signUp } from "./actions.js";
 
 type State = {
 	authenticatedUser: null | UserAuthResponseDto;
@@ -38,6 +38,18 @@ const { actions, name, reducer } = createSlice({
 			state.dataStatus = DataStatus.PENDING;
 		});
 		builder.addCase(signUp.rejected, (state) => {
+			state.authenticatedUser = null;
+			state.dataStatus = DataStatus.REJECTED;
+		});
+
+		builder.addCase(signIn.pending, (state) => {
+			state.dataStatus = DataStatus.PENDING;
+		});
+		builder.addCase(signIn.fulfilled, (state, action) => {
+			state.authenticatedUser = action.payload.data.user;
+			state.dataStatus = DataStatus.FULFILLED;
+		});
+		builder.addCase(signIn.rejected, (state) => {
 			state.authenticatedUser = null;
 			state.dataStatus = DataStatus.REJECTED;
 		});