+++ packages/shared/src/modules/auth/libs/enums/auth-api-path.enum.ts
@@ -1,5 +1,6 @@
 const AuthApiPath = {
 	AUTHENTICATED_USER: "/authenticated-user",
+	SIGN_IN: "/sign-in",
 	SIGN_UP: "/sign-up",
 } as const;