+++ packages/shared/src/modules/points-of-interest/libs/enums/points-of-interest-validation-message.enum.ts
@@ -1,0 +1,9 @@
+import { PointsOfInterestValidationRule } from "./points-of-interest-validation-rule.js";
+
+const PointsOfInterestValidationMessage = {
+	NAME_MAXIMUM_LENGTH: `Name must be less than ${String(PointsOfInterestValidationRule.NAME_MAXIMUM_LENGTH)} characters long.`,
+	NAME_MINIMUM_LENGTH: `Name must be at least ${String(PointsOfInterestValidationRule.NAME_MINIMUM_LENGTH)} characters long.`,
+	REQUIRED_FIELDS_FOR_UPDATE: "At least one field must be provided for update.",
+} as const;
+
+export { PointsOfInterestValidationMessage };