+++ packages/shared/src/modules/points-of-interest/libs/types/types.ts
@@ -1,0 +1,2 @@
+export { type PointsOfInterestRequestDto } from "./points-of-interest-request-dto.type.js";
+export { type PointsOfInterestResponseDto } from "./points-of-interest-response-dto.type.js";